Move visual map traversal to map component
