+++ src/Engine/GameEventNames.ts
@@ -1,0 +1,3 @@
+export const enum GameEventNames {
+    ChangeLocation = "ChangeLocation"
+}
