+++ src/Engine/Services/GameService.ts
@@ -23,6 +23,8 @@ import {getParsedDocument, InitEntityCollection} from "storyScript/EntityCreator
 import {IEquipment} from "storyScript/Interfaces/equipment.ts";
 import {ICombineResult} from "storyScript/Interfaces/combinations/combineResult.ts";
 import {ISoundService} from "storyScript/Interfaces/services/ISoundService.ts";
+import {gameEvents} from "storyScript/gameEvents.ts";
+import {GameEventNames} from "storyScript/GameEventNames.ts";
 
 export class GameService implements IGameService {
     constructor
@@ -112,7 +114,7 @@ export class GameService implements IGameService {
         }
 
         if (this._game.party?.currentLocationId) {
-            this._locationService.changeLocation(this._game.party.currentLocationId, false, this._game);
+            this._game.changeLocation(this._game.party.currentLocationId, false);
         }
     }
 
@@ -203,7 +205,7 @@ export class GameService implements IGameService {
             this._game.previousLocation = this._game.locations.get(previousLocationName);
         }
 
-        this._locationService.changeLocation(lastLocation.id, false, this._game);
+        this._game.changeLocation(lastLocation.id, false);
         this._game.state = GameState.Play;
     }
 
@@ -291,8 +293,13 @@ export class GameService implements IGameService {
         this.initCombinations();
         this._locationService.init();
 
+        gameEvents.register(GameEventNames.ChangeLocation, false);
+        
         this._game.changeLocation = (location, travel) => {
             this._locationService.changeLocation(location, travel, this._game);
+            
+            // Use a timeout here to give the Map component time to initialize before publishing the event.
+            setTimeout(() => gameEvents.publish(GameEventNames.ChangeLocation, { location, travel }));
 
             if (travel) {
                 this._dataService.saveGame(this._game);
