+++ src/Games/MyInteractiveMap/rules.ts
@@ -40,30 +40,6 @@ export function Rules(): IRules {
         },
 
         exploration: {
-            enterLocation(game: IGame, location: ICompiledLocation, travel?: boolean) {
-                const map = game.currentMap;
-                
-                setTimeout(() => {
-                    const coordString = map.locations.find(l => l.location === game.currentLocation.id)?.coords;
-
-                    if (coordString) {
-                        const coords = coordString.split(',');
-                        const coordLeft = parseInt(coords[0]);
-                        const coordTop = parseInt(coords[1]);
-                        const mapImage = <any>game.UIRootElement.getElementsByClassName('map-image')[0];
-                        const avatar = <any>game.UIRootElement.getElementsByClassName('avatar-image')[0];
-                        const mapMarginLeft = getMapMargin(mapImage, coordLeft, 'width');
-                        const mapMarginTop = getMapMargin(mapImage, coordTop, 'height');
-                        mapImage.style.marginLeft = `-${mapMarginLeft}px`;
-                        mapImage.style.marginTop = `-${mapMarginTop}px`;
-                        const avatarLeft = coordLeft - mapMarginLeft - avatar.width / 2;
-                        const avatarTop = coordTop - mapMarginTop - avatar.height / 2;
-                        avatar.style.left = `${avatarLeft}px`;
-                        avatar.style.top = `${avatarTop}px`;
-                    }  
-                    // This timeout is needed to allow the UI components to render and have the avatar dimensions available.
-                }, 250);
-            }
         },
 
         combat: {     
@@ -80,15 +56,4 @@ export function Rules(): IRules {
             }
         }
     };
-}
-
-function getMapMargin(mapImage: any, coord: number, dimension: string): number {
-    const mapContainer = mapImage.parentElement;
-    const clientDimension = 'client' + dimension.substring(0, 1).toUpperCase() + dimension.substring(1)
-    const mapDimension = mapContainer[clientDimension];
-    const viewPortCenter = mapDimension / 2;
-    const maxMargin = mapImage[dimension] - mapDimension;
-    let mapMargin = coord > viewPortCenter ? coord - viewPortCenter : 0;
-    mapMargin = mapMargin > maxMargin ? maxMargin : mapMargin;
-    return mapMargin;
 }
