+++ src/UI/Components/LocationMap/locationmap.component.html
@@ -1,5 +1,5 @@
 @let map = game.currentMap;
 <div id="map-container">
     <img class="map-image" src="{{ 'resources/' + map.mapImage }}" [attr.usemap]="'#' + map.id" alt="{{ map.name }}">
-    <img class="avatar-image" *ngIf="map.avatarImage" src="{{ 'resources/' + map.avatarImage }}" [attr.usemap]="'#' + map.id" alt="{{ map.name }}">
+    <img class="avatar-image" *ngIf="map.avatarImage" src="{{ 'resources/' + map.avatarImage }}" [attr.usemap]="'#' + map.id" alt="{{ map.name }}" style="display: none;">
 </div>
